move test to unit project
