+++ NotificationApi/NotificationApi.UnitTests/NotificationApi.UnitTests.csproj
@@ -13,6 +13,7 @@
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
     <PackageReference Include="Microsoft.TestPlatform" Version="16.6.1" />
     <PackageReference Include="Moq" Version="4.14.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="6.0.8" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="SonarAnalyzer.CSharp" Version="8.14.0.22654">
